Parse data from message, move output logic
This Lambda function needed to be triggered manually after an encoding job finished.  The lambda would list all of the current jobs, find the most recent finished job, and make transmuxings for that job.  This is not ideal because it could be susceptible to race conditions and other issues.  This PR enables the lambda to be triggered by a bitmovin notification that hits API Gateway and is relayed to our lambda.  The message triggers when a job has finished transfering to S3.  Our Lambda parses the job id and output url from the message, creates transmuxes for that job, and outputs the results to the proper destination.  Output creation logic was moved out of the create transmux function.  We only need one output per job, and we were getting one per transmux.  The transmux can share, and now they do.

Tags: bitmovin, notifications, callback