Update transmuxing parameters
We updated our encoding profile for new video encodings, so we needed to update which streams we want to transmux.  We decided to transmux the 720, smaller 1080, and 240p files as medium, high, and mobile.mp4 files respectively.  The files will act as a backup to the HLS streams.

tags: bitmovin, transmuxing, stream selection